backend/requirements.txt
backcall==0.1.0
decorator==4.4.0
Django==2.1
django-cors-headers==2.4.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.8.2
ipython==7.6.1
ipython-genutils==0.2.0
jedi==0.14.1
mongoengine==0.15.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pymongo==3.8.0
pytz==2019.1
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
